Changed Logic to implement printReceipt method. Implemented printReceipt method in CashRegister. Closes #13

+++ src/main/java/com/bridgephase/store/Product.java
@@ -23,7 +23,7 @@ public class Product {
 		this.name = name;
 	}
 
-	public double getWholesalePrice() {
+	public float getWholesalePrice() {
 		return wholesalePrice;
 	}
 
@@ -31,7 +31,7 @@ public class Product {
 		this.wholesalePrice = wholesalePrice;
 	}
 
-	public double getRetailPrice() {
+	public float getRetailPrice() {
 		return retailPrice;
 	}
